// src/Config.d.ts
export declare const CHUNK_SIZE = 32;
export declare const CHUNK_AREA: number;
export declare const CHUNK_VOLUME: number;
export declare const ISOLEVEL: number;
export declare enum Face {
    TOP = 0,
    BOTTOM = 1,
    FRONT = 2,
    BACK = 3,
    LEFT = 4,
    RIGHT = 5
}
//# sourceMappingURL=Config.d.ts.map